LayoutTests/imported/w3c/web-platform-tests/editing/other/editable-state-and-focus-in-shadow-dom-in-designMode.tentative-expected.txt

open

closed


PASS Waiting for load
FAIL Collapse selection into text in the open shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL Collapse selection into text in <div contenteditable> in the open shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL Set focus to <object> in the open shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL Set focus to <p tabindex="0"> in the open shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL SelectAll in the open shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL SelectAll in the <div contenteditable> in the open shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL Collapse selection into text in the closed shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL Collapse selection into text in <div contenteditable> in the closed shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL Set focus to <object> in the closed shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL Set focus to <p tabindex="0"> in the closed shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL SelectAll in the closed shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
FAIL SelectAll in the <div contenteditable> in the closed shadow DOM promise_test: Unhandled rejection with value: object "Error: Unknown source type "key"."
PASS Focus after Collapse selection into text in the open shadow DOM
FAIL Focus after Collapse selection into text in <div contenteditable> in the open shadow DOM assert_equals: <div contenteditable> should have focus after Collapse selection into text in <div contenteditable> in the open shadow DOM expected "div" but got ""
PASS Focus after Set focus to <object> in the open shadow DOM
FAIL Focus after Set focus to <p tabindex="0"> in the open shadow DOM assert_equals: The <p tabindex="0"> element should have focus after Set focus to <p tabindex="0"> in the open shadow DOM expected "p" but got ""
PASS Focus after Collapse selection into text in the closed shadow DOM
FAIL Focus after Collapse selection into text in <div contenteditable> in the closed shadow DOM assert_equals: <div contenteditable> should have focus after Collapse selection into text in <div contenteditable> in the closed shadow DOM expected "div" but got ""
PASS Focus after Set focus to <object> in the closed shadow DOM
FAIL Focus after Set focus to <p tabindex="0"> in the closed shadow DOM assert_equals: The <p tabindex="0"> element should have focus after Set focus to <p tabindex="0"> in the closed shadow DOM expected "p" but got ""
